+++ Library/Alliance/AllianceDonations.m
@@ -104,7 +104,7 @@
         NSString *r1 = rowData[@"club_name"];
         NSString *c1 = [[Globals i] numberFormat:rowData[@"currency_second"]];
         
-        return @{@"n1": [NSString stringWithFormat:@"%d", indexPath.row], @"r1": r1, @"c1": c1};
+        return @{@"n1": [NSString stringWithFormat:@"%ld", (long)indexPath.row], @"r1": r1, @"c1": c1};
     }
 }
